% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_lfc.R
\name{CenterMedian}
\alias{CenterMedian}
\title{Subtract the median of the given vector (for normalizing log2 fold changes).}
\usage{
CenterMedian(l)
}
\arguments{
\item{l}{Vector of effect sizes}
}
\value{
A vector of length 2 containing the two parameters
}
\description{
Subtract the median of the given vector (for normalizing log2 fold changes).
}
\seealso{
PsiLFC
}
