% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_lfc.R
\name{EmpiricalBayesPrior}
\alias{EmpiricalBayesPrior}
\title{Computes the prior parameters (i.e. pseudocounts incremented by 1) for
the log2 fold change distribution}
\usage{
EmpiricalBayesPrior(A, B, min.sd = 0)
}
\arguments{
\item{A}{Vector of counts from condition A}

\item{B}{Vector of counts from condition B}

\item{min.sd}{minimal standard deviation of the prior}
}
\value{
A vector of length 2 containing the two parameters
}
\description{
Computes the prior parameters (i.e. pseudocounts incremented by 1) for
the log2 fold change distribution
}
\examples{
  EmpiricalBayesPrior(rnorm(1000,200),rnorm(1000,100))
}
\seealso{
PsiLFC
}
