% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfc_basics.R
\name{ltop}
\alias{ltop}
\title{Inverse logit transformation to obtain proportion representation from the
log fold change representation.}
\usage{
ltop(l)
}
\arguments{
\item{l}{Effect size in log2 fold change representation}
}
\value{
The proportion representation of the effect size
}
\description{
Inverse logit transformation to obtain proportion representation from the
log fold change representation.
}
\examples{
ptol(0)
ptol(1)
}
\seealso{
ptol

Other Effect size representations: \code{\link{ptol}}
}
