% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_lfc.R
\name{NormLFC}
\alias{NormLFC}
\title{Standard LFC effect size estimator}
\usage{
NormLFC(A, B, pseudo = c(1, 1), normalizeFun = CenterMedian)
}
\arguments{
\item{A}{Vector of counts from condition A}

\item{B}{Vector of counts from condition B}

\item{pseudo}{Vector of length 2 of the pseudo counts}

\item{normalizeFun}{Function to normalize the obtained effect sizes}
}
\value{
The vector containing the estimates
}
\description{
Computes the standard, normalized log2 fold change with given pseudocounts
}
\examples{
  NormLFC(rnorm(1000,200),rnorm(1000,100))
}
