% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_lfc.R
\name{PsiLFC}
\alias{PsiLFC}
\title{Computes the optimal effect size estimate.}
\usage{
PsiLFC(A, B, prior = EmpiricalBayesPrior(A, B), normalizeFun = CenterMedian)
}
\arguments{
\item{A}{Vector of counts from condition A}

\item{B}{Vector of counts from condition B}

\item{prior}{Vector of length 2 of the prior parameters}

\item{normalizeFun}{Function to normalize the obtained effect sizes}
}
\value{
A vector of length 2 containing the two parameters
}
\description{
Computes the optimal effect size estimate.
}
