% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfc_basics.R
\name{ptol}
\alias{ptol}
\title{Logit transformation to obtain the log fold change representation from the
proportion representation.}
\usage{
ptol(p)
}
\arguments{
\item{p}{Effect size in proportion representation}
}
\value{
The log2 fold change representation of the effect size
}
\description{
Logit transformation to obtain the log fold change representation from the
proportion representation.
}
\examples{
ptol(0.5)
ptol(2/3)
}
\seealso{
ltop

Other Effect size representations: \code{\link{ltop}}
}
