% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_lfc.R
\name{results}
\alias{results}
\title{Psi LFC effect size estimator for DESeq2}
\usage{
results(object, contrast, cre = F, ...)
}
\arguments{
\item{object}{the DESeq2DataSet object}

\item{contrast}{Vector of length 3, specifying the variable and the two levels
to compute effect sizes for (<name>,<A>,<B>)}

\item{cre}{Compute credible intervals as well? (can also be a vector of quantiles)}
}
\value{
Either a vector containing the estimates, or a data frame containing
    the credible interval as well
}
\description{
Drop-in replacement for DESeq2's results function for simple settings involving
a single variable. Appends the PsiLFC estimate.
}
